refactor: extract native driver route authority (#1728)
